Extract helpers into xtra-libp2p::tests::util module
